Change how codecs are chosen for audio
The codecs are now chosen through table-driven code rather than a
sequence of conditionals.  Codecs which the browser supports "probably"
are now preferred over ones which the browser supports "maybe".  For
browsers that support both Vorbis and MPEG audio, Vorbis is now
preferred instead of MPEG (this only applies to Chrome).